+++ framework/rbac/migrations/schema-mssql.sql
@@ -65,37 +65,3 @@ create table [auth_assignment]
 
 create index [auth_assignment_user_id_idx] on [auth_assignment] ([user_id]);
 
-CREATE TRIGGER dbo.trigger_auth_item_child
-    ON dbo.[auth_item]
-    INSTEAD OF DELETE, UPDATE
-    AS
-    DECLARE @old_name VARCHAR (64) = (SELECT name FROM deleted)
-    DECLARE @new_name VARCHAR (64) = (SELECT name FROM inserted)
-    BEGIN
-        IF COLUMNS_UPDATED() > 0
-        BEGIN
-            IF @old_name <> @new_name
-                BEGIN
-                    ALTER TABLE auth_item_child NOCHECK CONSTRAINT FK__auth_item__child;
-                    UPDATE auth_item_child SET child = @new_name WHERE child = @old_name;
-                END
-                UPDATE auth_item
-                SET name = (SELECT name FROM inserted),
-                type = (SELECT type FROM inserted),
-                description = (SELECT description FROM inserted),
-                rule_name = (SELECT rule_name FROM inserted),
-                data = (SELECT data FROM inserted),
-                created_at = (SELECT created_at FROM inserted),
-                updated_at = (SELECT updated_at FROM inserted)
-                WHERE name IN (SELECT name FROM deleted)
-                IF @old_name <> @new_name
-                    BEGIN
-                        ALTER TABLE auth_item_child CHECK CONSTRAINT FK__auth_item__child;
-                    END
-                END
-                ELSE
-                    BEGIN
-                        DELETE FROM dbo.[auth_item_child] WHERE parent IN (SELECT name FROM deleted) OR child IN (SELECT name FROM deleted);
-                        DELETE FROM dbo.[auth_item] WHERE name IN (SELECT name FROM deleted);
-                    END
-        END;